+++ app/contact/page.tsx
@@ -118,7 +118,7 @@ const Contact = () => {
                     </div>
                 </div>
             </div>
-            <Footer />
+        
         </ Container>
